chore: Move the filtering logic to a separate file
This takes advantage of generic functions to reuse the same
logic for both vault types.
